# docs/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Honeytree API
  version: 1.0.0
  description: |
    Contract between the Honeytree API (Agent 1), the web app (Agent 2) and the
    media module (Agent 3). All paths are prefixed with `/api/v1`.

    ## Conventions
    - **Auth**: httpOnly session cookie `ht_session` (set by register/login). Endpoints marked
      `security: [cookieAuth]` need it (401 `UNAUTHENTICATED` otherwise). Others accept it optionally,
      and use it to fill `likedByMe`.
    - **CSRF**: the API rejects non-GET requests whose `Origin` header is not an allowed web origin.
      Send requests with `credentials: 'include'`.
    - **Errors**: `{ "error": { "code": "STRING", "message": "…", "details": … } }`.
    - **Pagination**: cursor based. `?cursor=&limit=` → `{ items, nextCursor }` (`nextCursor` is `null` on the last page).
      Treat cursors as opaque strings.
    - **`likedByMe`** is present on every game and comment (`false` when anonymous).
    - **`karmaAwarded`** is returned by every write endpoint that can earn or lose karma. Positive = earned, negative =
      reversed, `0` = nothing (own content, daily cap, text too short). Use it for the "+1 karma 🍯" toast.
    - **Text** (`body`, `description`, …) is plain Markdown. The API strips raw HTML on input; the web app must render
      Markdown with HTML disabled.
    - **Media URLs** are absolute and public. Build files have `url: null`; use `GET /games/{id}/download` (Agent 3).
...

...

...
servers:
...
security: []
...
paths:
...
  /users/me:
    patch:
      tags: [Users]
      security: [{ cookieAuth: [] }]
      summary: Edit own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { type: string, maxLength: 40 }
                bio: { type: string, maxLength: 500 }
                links: { type: array, maxItems: 5, items: { $ref: '#/components/schemas/Link' } }
                avatarMediaId: { type: string, format: uuid, nullable: true, description: 'A `ready` media of kind `avatar` owned by you' }
      responses:
        '200': { description: 'OK', content: { application/json: { schema: { type: object, properties: { user: { $ref: '#/components/schemas/UserProfile' } } } } } }
        '422': { $ref: '#/components/responses/Validation' }
  /users/{username}:
    get:
      tags: [Users]
      summary: Profile with stats
      parameters: [{ $ref: '#/components/parameters/Username' }]
      responses:
        '200': { description: 'OK', content: { application/json: { schema: { type: object, properties: { user: { $ref: '#/components/schemas/UserProfile' } } } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/{username}/games:
    get:
      tags: [Users]
      summary: Games of a user
      description: Published games only, except when `status` is requested by the user themself.
      parameters:
        - { $ref: '#/components/parameters/Username' }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { name: status, in: query, schema: { type: string, enum: [published, draft, unpublished, all], default: published }, description: 'Non-published values are honoured only for your own profile' }
      responses:
        '200': { description: 'OK', content: { application/json: { schema: { $ref: '#/components/schemas/GamePage' } } } }
  /users/{username}/reviews:
    get:
      tags: [Users]
      summary: Reviews written by a user
      parameters:
        - { $ref: '#/components/parameters/Username' }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/ReviewWithGame' } }
                  nextCursor: { type: string, nullable: true }
  /users/{username}/activity:
    get:
      tags: [Users]
      summary: Recent public activity (last 30 items, no pagination)
      parameters: [{ $ref: '#/components/parameters/Username' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/ActivityItem' } }
...
  /games/{id}/reviews:
    get:
      summary: List reviews with rating summary
      tags: [Social]
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Review' } }
                  nextCursor: { type: string, nullable: true }
                  summary: { $ref: '#/components/schemas/RatingSummary' }
    post:
      tags: [Social]
      security: [{ cookieAuth: [] }]
      summary: Review a game (one per user; creators cannot review their own game)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                body: { type: string, maxLength: 4000, description: 'Karma is earned only when ≥ 20 characters' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  review: { $ref: '#/components/schemas/Review' }
                  karmaAwarded: { type: integer }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /reviews/{id}:
    patch:
      summary: Edit own review
      tags: [Social]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                body: { type: string, maxLength: 4000 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  review: { $ref: '#/components/schemas/Review' }
                  karmaAwarded: { type: integer }
    delete:
      summary: Delete a review (author or admin)
      tags: [Social]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: 'Deleted', content: { application/json: { schema: { type: object, properties: { karmaAwarded: { type: integer } } } } } }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: ht_session }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    Username: { name: username, in: path, required: true, schema: { type: string } }
    Cursor: { name: cursor, in: query, schema: { type: string }, description: 'Opaque cursor from `nextCursor`' }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
  responses:
...
    Forbidden: { description: 'Not allowed (`FORBIDDEN`, `SELF_ACTION`)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: 'Not found', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Conflict: { description: 'Already exists', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Validation: { description: 'Invalid input (`VALIDATION_ERROR`, `PUBLISH_REQUIREMENTS`)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details: {}
    Link:
      type: object
      required: [label, url]
      properties:
        label: { type: string, maxLength: 30 }
        url: { type: string, format: uri }
    UserSummary:
      type: object
      required: [id, username, displayName, avatarUrl]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string, nullable: true }
    UserStats:
      type: object
      properties:
        likesReceived: { type: integer, description: 'Likes on the user''s games + comments' }
        gamesCount: { type: integer, description: 'Published games' }
        ratingAvg: { type: number, nullable: true, description: 'Average rating of all reviews received on the user''s games (null when none)' }
        ratingCount: { type: integer }
        karma: { type: integer }
    UserProfile:
      allOf:
        - { $ref: '#/components/schemas/UserSummary' }
        - type: object
          properties:
            bio: { type: string }
            links: { type: array, items: { $ref: '#/components/schemas/Link' } }
            createdAt: { type: string, format: date-time }
            stats: { $ref: '#/components/schemas/UserStats' }
...
    GameSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        title: { type: string }
        shortDescription: { type: string }
        tags: { type: array, items: { type: string } }
        platforms: { type: array, items: { type: string, enum: [windows, mac, linux, web, android] } }
        coverUrl: { type: string, nullable: true, description: '640px card variant when available' }
        owner: { $ref: '#/components/schemas/UserSummary' }
        status: { type: string, enum: [draft, published, unpublished, removed] }
        likesCount: { type: integer }
        downloadsCount: { type: integer }
        commentsCount: { type: integer }
        reviewsCount: { type: integer }
        ratingAvg: { type: number, nullable: true }
        publishedAt: { type: string, format: date-time, nullable: true }
        likedByMe: { type: boolean }
...
    GamePage:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/GameSummary' } }
        nextCursor: { type: string, nullable: true }
    Review:
      type: object
      properties:
        id: { type: string, format: uuid }
        gameId: { type: string, format: uuid }
        user: { $ref: '#/components/schemas/UserSummary' }
        rating: { type: integer, minimum: 1, maximum: 5 }
        body: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ReviewWithGame:
      allOf:
        - { $ref: '#/components/schemas/Review' }
        - type: object
          properties:
            game: { type: object, properties: { id: { type: string }, slug: { type: string }, title: { type: string }, coverUrl: { type: string, nullable: true } } }
    RatingSummary:
      type: object
      properties:
        ratingAvg: { type: number, nullable: true }
        ratingCount: { type: integer }
        distribution:
          type: object
          description: 'Number of reviews per star: keys "1".."5"'
          additionalProperties: { type: integer }
...
    ActivityItem:
      type: object
      properties:
        type: { type: string, enum: [comment, suggestion, review, like, published] }
        createdAt: { type: string, format: date-time }
        excerpt: { type: string, nullable: true }
        rating: { type: integer, nullable: true }
        game: { type: object, properties: { id: { type: string }, slug: { type: string }, title: { type: string } } }